wip commit: removing side effects from tokens reducer to actions
