+++ src/logic/tokens/store/actions/fetchTokens.js
@@ -9,7 +9,7 @@ import { type GlobalState } from '~/store/index'
 import { makeToken, type Token, type TokenProps } from '~/logic/tokens/store/model/token'
 import { ensureOnce } from '~/utils/singleton'
 import { getActiveTokenAddresses, getTokens } from '~/logic/tokens/utils/activeTokensStorage'
-import { getSafeEthToken } from '~/utils/tokens'
+import { getSafeEthToken } from '~/logic/tokens/utils/tokenHelpers'
 import { enhancedFetch } from '~/utils/fetch'
 import addTokens from './addTokens'
 import { getRelayUrl } from '~/config/index'
