+++ src/logic/tokens/store/reducer/tokens.js
@@ -1,5 +1,5 @@
 // @flow
-import { List, Map } from 'immutable'
+import { Map } from 'immutable'
 import { handleActions, type ActionType } from 'redux-actions'
 import addToken, { ADD_TOKEN } from '~/logic/tokens/store/actions/addToken'
 import removeToken, { REMOVE_TOKEN } from '~/logic/tokens/store/actions/removeToken'
@@ -10,31 +10,16 @@ import enableToken, { ENABLE_TOKEN } from '~/logic/tokens/store/actions/enableTo
 import {
   setActiveTokenAddresses,
   getActiveTokenAddresses,
-  setToken,
-  removeTokenFromStorage,
 } from '~/logic/tokens/utils/activeTokensStorage'
-import { ensureOnce } from '~/utils/singleton'
-import { calculateActiveErc20TokensFrom } from '~/utils/tokens'
 
 export const TOKEN_REDUCER_ID = 'tokens'
 
 export type State = Map<string, Map<string, Token>>
 
-const setTokensOnce = ensureOnce(setActiveTokenAddresses)
-
-const removeFromActiveTokens = (safeAddress: string, tokenAddress: string) => {
-  const activeTokens = getActiveTokenAddresses(safeAddress)
-  const index = activeTokens.indexOf(tokenAddress)
-  setActiveTokenAddresses(safeAddress, activeTokens.delete(index))
-}
-
 export default handleActions(
   {
     [ADD_TOKENS]: (state: State, action: ActionType<typeof addTokens>): State => {
       const { safeAddress, tokens } = action.payload
-
-      const activeAddresses: List<Token> = calculateActiveErc20TokensFrom(tokens.toList())
-      setTokensOnce(safeAddress, activeAddresses)
 
       return state.update(safeAddress, (prevSafe: Map<string, Token>) => {
         if (!prevSafe) {
@@ -48,17 +33,12 @@ export default handleActions(
       const { safeAddress, token } = action.payload
 
       const tokenAddress = token.get('address')
-      const activeTokens = getActiveTokenAddresses(safeAddress)
-      setActiveTokenAddresses(safeAddress, activeTokens.push(tokenAddress))
-      setToken(safeAddress, token)
       return state.setIn([safeAddress, tokenAddress], token)
     },
     [REMOVE_TOKEN]: (state: State, action: ActionType<typeof removeToken>): State => {
       const { safeAddress, token } = action.payload
 
       const tokenAddress = token.get('address')
-      removeFromActiveTokens(safeAddress, tokenAddress)
-      removeTokenFromStorage(safeAddress, token)
       return state.removeIn([safeAddress, tokenAddress])
     },
     [DISABLE_TOKEN]: (state: State, action: ActionType<typeof disableToken>): State => {
