+++ src/routes/safe/component/SendToken/index.jsx
@@ -7,7 +7,7 @@ import { sleep } from '~/utils/timer'
 import { type Safe } from '~/routes/safe/store/model/safe'
 import { getStandardTokenContract } from '~/logic/tokens/store/actions/fetchTokens'
 import { type Token } from '~/logic/tokens/store/model/token'
-import { isEther } from '~/utils/tokens'
+import { isEther } from '~/logic/tokens/utils/tokens'
 import { EMPTY_DATA } from '~/logic/wallets/ethTransactions'
 import { toNative } from '~/logic/wallets/tokens'
 import { createTransaction, getSafeEthereumInstance } from '~/logic/safe/safeFrontendOperations'
